server/requirements.txt
grpcio
grpcio-tools
Flask
requests
psutil
